android/android-grpc/python/aemu-grpc/pyproject.toml
[build-system]
# AVOID CHANGING REQUIRES: IT WILL BE UPDATED BY PYSCAFFOLD!
requires = ["setuptools>=46.1.0", "wheel", "grpcio-tools==1.48.1"]
build-backend = "setuptools.build_meta"

# [tool.setuptools_scm]
# # For smarter version schemes and other configuration options,
# # check out https://github.com/pypa/setuptools_scm
# version_scheme = "no-guess-dev"
[project]
name = "aemu-grpc"
authors = [{ name = "Erwin Jansen", email = "jansene@google.com" }]
license = { file = "LICENSE.TXT" }
readme = "README.md"
version = "0.0.1"
classifiers = ["Topic :: Software Development"]


dependencies = [
    "grpcio==1.48.1",
    "protobuf==3.20.1",
    "tink==1.7.0; sys_platform != 'win32'",
    "importlib-metadata; python_version<'3.8'",
]

[project.optional-dependencies]
tests = [
    "pytest==7.1.3",
    "pytest-mock",
    "mock==4.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]
tests = ["pytest==7.1.3", "pytest_mock==3.8.2"]